feat: Seperated trip results and details
